fix: move viewport height to graph page only
